+++ src/dlls/objects/399_ECSH_Shrine/ECSH_Shrine.c
@@ -223,9 +223,9 @@ void ecshShrine_updateHoverMotion(GameObject* obj) {
         return;
     }
 
-    state->orbitPhaseA = (s16)(state->orbitPhaseA + (s32)(ECSH_SHRINE_ORBIT_RATE_A * timeDelta));
-    state->orbitPhaseB = (s16)(state->orbitPhaseB + (s32)(ECSH_SHRINE_ORBIT_RATE_B * timeDelta));
-    state->orbitPhaseC = (s16)(state->orbitPhaseC + (s32)(ECSH_SHRINE_ORBIT_RATE_C * timeDelta));
+    state->orbitPhaseA = state->orbitPhaseA + (s32)(ECSH_SHRINE_ORBIT_RATE_A * timeDelta);
+    state->orbitPhaseB = state->orbitPhaseB + (s32)(ECSH_SHRINE_ORBIT_RATE_B * timeDelta);
+    state->orbitPhaseC = state->orbitPhaseC + (s32)(ECSH_SHRINE_ORBIT_RATE_C * timeDelta);
 
     obj->anim.localPosY =
         ECSH_SHRINE_ORBIT_HEIGHT +
